audio: Fix media server crash at the start of voice call
Media server is crashing at the start of voice call. This is due
to no strategy is supported for AUDIO_STREAM_PATCH stream type.
AUDIO_STREAM_PATCH stream is added as part of L-MR1 upgrade
and it's for internal flinger tracks. Fix this issue by skipping
handleIncallSonification() for AUDIO_STREAM_PATCH stream.

Change-Id: Id12331854ee869e697e71c351a411c75d169e16b