audiopolicy: fix for voice call to work on USB headset during concurrency
When an input stream is active on USB headset and if a voice call
is received, the proxy_open() for voice call fails and the screen freezes.
All active inputs must be closed before opening input stream for
voice call and also all new requests to open input stream must be
blocked.

Change-Id: I1fd966b4f60a25f890fe35251fa127868cec2cdd
CRs-fixed: 876993